continue refactoring, keep shrinking lines of code
