+++ pcs/src/lib.rs
@@ -1,6 +1,8 @@
 mod traits;
 pub use traits::PolynomialCommitmentScheme;
 
+pub const PCS_SOUNDNESS_BITS: usize = 128;
+
 pub mod raw;
 
 pub mod orion;
